protection controller et ajour popup panier

+++ controller/controllerCategorie.php
@@ -19,41 +19,77 @@ class controllerCategorie
 
     public static function create()
     {
-        $view = 'form';
-        $name = 'created';
-        $numCategorie = '';
-        require File::build_path(array('view', 'view.php'));
+        if (isset($_SESSION['login']) && $_SESSION['isAdmin']=='1') {
+            $view = 'form';
+            $name = 'created';
+            $numCategorie = '';
+            require File::build_path(array('view', 'view.php'));
+        } else if (isset($_SESSION['login'])) {
+            echo '<p class="alert alert-danger">Vous n\'avez pas la permission de réaliser cela !</p>';
+            controllerBook::readAll();
+        } else {
+            ControllerUtilisateur::login();
+        }
     }
 
     public static function delete()
     {
-        ModelCategorie::delete();
-        self::readAll();
+        if (isset($_SESSION['login']) && $_SESSION['isAdmin']=='1') {
+            ModelCategorie::delete();
+            self::readAll();
+        } else if (isset($_SESSION['login'])) {
+            echo '<p class="alert alert-danger">Vous n\'avez pas la permission de réaliser cela !</p>';
+            controllerBook::readAll();
+        } else {
+            ControllerUtilisateur::login();
+        }
     }
 
     public static function created()
     {
-        $data = array('nomCategorie' => $_POST['nomCategorie']);
-        ModelCategorie::saveGen($data);
-        self::readAll();
+        if (isset($_SESSION['login']) && $_SESSION['isAdmin']=='1') {
+            $data = array('nomCategorie' => $_POST['nomCategorie']);
+            ModelCategorie::saveGen($data);
+            self::readAll();
+            echo "<div class='alert alert-success'>La catégorie a bien été créée ! </div>";
+        } else if (isset($_SESSION['login'])) {
+            echo '<p class="alert alert-danger">Vous n\'avez pas la permission de réaliser cela !</p>';
+            controllerBook::readAll();
+        } else {
+            ControllerUtilisateur::login();
+        }
     }
 
     public static function update() {
-        $view = 'form';
-        $name = 'updated';
-        $numCategorie = $_GET['numCategorie'];
-        $categorie = ModelCategorie::select($numCategorie)[0];
-        require File::build_path(array('view', 'view.php'));
+        if (isset($_SESSION['login']) && $_SESSION['isAdmin']=='1') {
+            $view = 'form';
+            $name = 'updated';
+            $numCategorie = $_GET['numCategorie'];
+            $categorie = ModelCategorie::select($numCategorie)[0];
+            require File::build_path(array('view', 'view.php'));
+        } else if (isset($_SESSION['login'])) {
+            echo '<p class="alert alert-danger">Vous n\'avez pas la permission de réaliser cela !</p>';
+            controllerBook::readAll();
+        } else {
+            ControllerUtilisateur::login();
+        }
     }
 
     public static function updated() {
-        $numCategorie = $_POST['numCategorie'];
-        $nomCategorie = $_POST['nomCategorie'];
-        $data = array('numCategorie' => $numCategorie,
-            'nomCategorie' => $nomCategorie);
-        modelCategorie::update($data);
+        if (isset($_SESSION['login']) && $_SESSION['isAdmin']=='1') {
+            $numCategorie = $_POST['numCategorie'];
+            $nomCategorie = $_POST['nomCategorie'];
+            $data = array('numCategorie' => $numCategorie,
+                'nomCategorie' => $nomCategorie);
+            modelCategorie::update($data);
 
-        self::readAll();
-        echo "La categorie a bien été modifié !";
+            self::readAll();
+            echo "<div class='alert alert-success'>La catégorie a bien été modifiée ! </div>";
+        } else if (isset($_SESSION['login'])) {
+            echo '<p class="alert alert-danger">Vous n\'avez pas la permission de réaliser cela !</p>';
+            controllerBook::readAll();
+        } else {
+            ControllerUtilisateur::login();
+        }
     }
 }
